+++ scrolly/_cli/_introspect/__init__.py
@@ -26,6 +26,9 @@ import click
 from scrolly._cli._introspect._assets import assets_command
 from scrolly._cli._introspect._elements import elements_command
 from scrolly._cli._introspect._slides import slides_command
+from scrolly._cli._introspect._snaps import snaps_command
+from scrolly._cli._introspect._snapshot import snapshot_command
+from scrolly._cli._introspect._timeline import timeline_command
 
 
 @click.group(name="introspect")
@@ -35,4 +38,7 @@ def introspect() -> None:
 
 introspect.add_command(slides_command)
 introspect.add_command(elements_command)
+introspect.add_command(snaps_command)
+introspect.add_command(timeline_command)
+introspect.add_command(snapshot_command)
 introspect.add_command(assets_command)